Finished
Everything cleaned up, essential features are added, not complete
