+++ SimpleFinance/GroupCell.swift
@@ -18,19 +18,10 @@ class GroupCell: UITableViewCell {
         return label
         
     }()
-    
-    let colorLabel: UILabel = {
-        
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        return label
-        
-    }()
 
     override init(style: UITableViewCellStyle, reuseIdentifier: String?) {
         super.init(style: .subtitle, reuseIdentifier: "groupCell")
         
-        addSubview(colorLabel)
         addSubview(progressLabel)
        
         progressLabel.rightAnchor.constraint(equalTo: self.rightAnchor).isActive = true
@@ -38,14 +29,6 @@ class GroupCell: UITableViewCell {
         progressLabel.widthAnchor.constraint(equalToConstant: 120).isActive = true
         progressLabel.heightAnchor.constraint(equalTo: (textLabel?.heightAnchor)!).isActive = true
         progressLabel.textColor = UIColor.lightGray
-        
-        colorLabel.leftAnchor.constraint(equalTo: self.leftAnchor).isActive = true
-        colorLabel.centerYAnchor.constraint(equalTo: self.topAnchor, constant: 20).isActive = true
-        colorLabel.widthAnchor.constraint(equalToConstant: 120).isActive = true
-        colorLabel.heightAnchor.constraint(equalTo: (textLabel?.heightAnchor)!).isActive = true
-        colorLabel.backgroundColor = UIColor.yellow
-        colorLabel.alpha = 0.5
-        
         
         
     }
